Port the test for dst folder exists
