+++ sql/hive/src/test/scala/org/apache/spark/sql/hive/execution/command/RenameTableSuite.scala
@@ -22,6 +22,4 @@ import org.apache.spark.sql.execution.command.v1
 /**
  * The class contains tests for the `RENAME TABLE` command to check V1 Hive external table catalog.
  */
-class RenameTableSuite extends v1.RenameTableSuiteBase with CommandSuiteBase {
-
-}
+class RenameTableSuite extends v1.RenameTableSuiteBase with CommandSuiteBase
